exits now

+++ Scheduler.cpp
@@ -13,6 +13,7 @@ Scheduler::Scheduler() {
   totalJobs = 0; // starts at the first line of a file
   totalJobsToDo = 100;
   fin.open("Input.txt");
+  exitScheduler = false;
 }
 Scheduler::Scheduler(int numJobs, int procs) {
   avaliableProcs = procs;
@@ -20,13 +21,19 @@ Scheduler::Scheduler(int numJobs, int procs) {
   totalJobs = 0;
   totalJobsToDo = numJobs;
   fin.open("Input.txt");
+  exitScheduler = false;
 }
 
 //runs jobs from file w/o user input
 void Scheduler::Run() {
   while (!fin.eof() || !procaQueue.empty() || !running.empty()){  
     Tick();
+      if(exitScheduler){
+      std::cout << "exiting" <<std::endl;
+      return;
+      }
     decrementTJobs();
+
   }
 }
 
@@ -52,10 +59,16 @@ void Scheduler::Tick() {
  // waitForUserInput();
  if (!fin.eof())
  {
-  getAJobFromTextFile();
+  exitScheduler = getAJobFromTextFile();
  }
+ if(!procaQueue.empty() || !exitScheduler)
+ {
   scheduleJob();
+ }
+ if(!running.empty() || !exitScheduler)
+ {
   deleteByTimer();
+ }
 }
 
 /*
@@ -139,23 +152,21 @@ bool Scheduler::getAJobFromTextFile() {
   string desc;
   int procs = 0;    
   int ticks = 0;
-  char eatNewLine; 
   getline(fin, desc, ' ');
   for(int i = 0; i < desc.length(); i++)
   {
     desc[i] = tolower(desc[i]);
   }
-
-  if(desc.compare("exit"))
+  if(desc.find("exit") == string::npos)
   {
-  std::cout<<"here" << std::endl;
+    std::cout<< "here" << std::endl;
     fin >> procs;
     fin >> ticks;
     fin.ignore(1000, '\n');
     insertJob(totalJobs++, procs, ticks, desc);
-    return true;
+    return false;
   }
-  return false;
+  return true;
   
 }
